% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_auto_variogram.R
\name{plot_auto_variogram}
\alias{plot_auto_variogram}
\title{Plot empirical and fitted variogram}
\usage{
plot_auto_variogram(pts, vgm_model, cutoff = 3000, width = 100)
}
\arguments{
\item{pts}{A sf object or data.frame.}

\item{vgm_model}{A gstat variogram model object.}

\item{cutoff}{Maximum distance for the empirical variogram.}

\item{width}{Bin width for the empirical variogram.}
}
\description{
Plots the empirical variogram together with the fitted model.
}
