% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{run_sim}
\alias{run_sim}
\title{Run Conditional Gaussian Simulations (SGS)}
\usage{
run_sim(
  gdata,
  nsim,
  newdata,
  vmodel,
  nmin = 0,
  nmax = 0,
  distmax = Inf,
  ncores = NULL,
  debug.level = -1,
  write_to_grass = TRUE,
  grass_prefix = "sgs_sim"
)
}
\arguments{
\item{gdata}{sf or Spatial object: observed points}

\item{nsim}{integer: number of simulations}

\item{newdata}{sf or Spatial object: prediction grid}

\item{vmodel}{variogram model from gstat::fit.variogram()}

\item{nmin}{optional kriging parameter Minimum number of neighbors}

\item{nmax}{optional kriging parameter Maximum number of neighbors}

\item{distmax}{optional kriging parameter Maximum search distance}

\item{ncores}{number of parallel workers}

\item{debug.level}{integer controlling kriging verbosity}

\item{write_to_grass}{logical: whether to save results to GRASS}

\item{grass_prefix}{character: GRASS raster name prefix}
}
\value{
A list of simulated realizations
}
\description{
Perform parallel conditional Gaussian simulations using gstat and save results into GRASS.
}
