% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_auto_variogram.R
\name{plot_auto_variogram}
\alias{plot_auto_variogram}
\title{Plot empirical and fitted variogram}
\usage{
plot_auto_variogram(vg, vgm_model)
}
\arguments{
\item{vg}{A gstat::variogram object.}

\item{vgm_model}{A gstat variogram model object.}
}
\value{
A plot of the empirical variogram with the fitted model.
}
\description{
Plots the empirical variogram together with the fitted model.
}
